+++ ogma-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for ogma-core
 
+## [1.8.0] - 2025-07-13
+
+* Version bump 1.8.0 (#275).
+* Introduce overview command (#272).
+* Extend code backends to accept expressions to monitor as arguments (#121).
+
 ## [1.7.0] - 2025-03-21
 
 * Version bump 1.7.0 (#268).